Enable Keystone V3 Authentication
Move auth into the tools_common module.
port_cleanup now ksv3 capable.
